requirements.txt
discord.py==2.3.2
python-decouple==3.8
youtube_search_python==1.6.6
yt_dlp==2023.12.30
